Replacing the Ninja api with Disease api url, closes #29
