+++ src/services/API/Disease.ts
@@ -1,5 +1,16 @@
 import axios from "axios";
 
+export type GlobalCases = {
+  confirmed: number;
+  deaths: number;
+  recovered: number;
+  deathsToday: number;
+  newConfirmed: number;
+  totalSerious: number;
+};
+
+export type CountryCases = GlobalCases;
+
 export type TimelineData = {
   date: string;
   active: number;
@@ -13,6 +24,58 @@ export type CasesTimeline = {
   cases: any;
   deaths: any;
   recovered: any;
+};
+
+export const fetchGlobalData: () => Promise<GlobalCases> = () => {
+  return new Promise((resolve, reject) => {
+    axios
+      .get("https://disease.sh/v3/covid-19/all")
+      .then((response) => {
+        if (response.data.cases) {
+          const results: GlobalCases = {
+            confirmed: response.data.cases,
+            deaths: response.data.deaths,
+            recovered: response.data.recovered,
+            deathsToday: response.data.todayDeaths,
+            newConfirmed: response.data.todayCases,
+            totalSerious: response.data.critical,
+          };
+          return resolve(results);
+        }
+        reject(new Error("Ninja API returned no data"));
+      })
+      .catch((err) => {
+        console.log("Ninja API error (Global): ", err.message);
+        reject(err.message);
+      });
+  });
+};
+
+export const fetchCountryData: (country: string) => Promise<CountryCases> = (
+  country
+) => {
+  return new Promise((resolve, reject) => {
+    axios
+      .get(`https://disease.sh/v3/covid-19/countries/${country}?strict=true`)
+      .then((response) => {
+        if (response.data.cases) {
+          const results: CountryCases = {
+            confirmed: response.data.cases,
+            deaths: response.data.deaths,
+            recovered: response.data.recovered,
+            deathsToday: response.data.todayDeaths,
+            newConfirmed: response.data.todayCases,
+            totalSerious: response.data.critical,
+          };
+          return resolve(results);
+        }
+        reject(new Error("Ninja API returned no data!"));
+      })
+      .catch((err) => {
+        console.log("Ninja API error (Country): ", err.message);
+        reject(err.message);
+      });
+  });
 };
 
 // get cases timeline
